requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
jwt==1.3.1
PyJWT==2.8.0
groq==0.4.1